Move block contents to datastructures package
